@@ -25,20 +25,20 @@ dependencies:
     sdk: flutter
   pluto_grid:
     path: ../
-  faker: ^2.0.0
-  url_launcher: ^6.1.6
-  font_awesome_flutter: ^10.2.1
+  faker: ^2.2.0
+  url_launcher: ^6.3.2
+  font_awesome_flutter: ^10.12.0
   rainbow_color: ^2.0.1
-  pluto_menu_bar: ^2.0.1
+  pluto_menu_bar: ^3.0.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.4
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
