@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:faker/faker.dart';
+import 'package:faker/faker.dart' hide Color;
 import 'package:flutter/material.dart';
 import 'package:pluto_grid/pluto_grid.dart';
 
@@ -10,7 +10,7 @@ import '../../widget/pluto_example_screen.dart';
 class AddAndRemoveColumnRowScreen extends StatefulWidget {
   static const routeName = 'add-and-remove-column-row';
 
-  const AddAndRemoveColumnRowScreen({Key? key}) : super(key: key);
+  const AddAndRemoveColumnRowScreen({super.key});
 
   @override
   _AddAndRemoveColumnRowScreenState createState() =>
@@ -167,8 +167,7 @@ class _AddAndRemoveColumnRowScreenState
 class _Header extends StatefulWidget {
   const _Header({
     required this.stateManager,
-    Key? key,
-  }) : super(key: key);
+  });
 
   final PlutoGridStateManager stateManager;
 
