@@ -7,7 +7,7 @@ import '../../widget/pluto_example_screen.dart';
 class ColumnSortingScreen extends StatefulWidget {
   static const routeName = 'feature/column-sorting';
 
-  const ColumnSortingScreen({Key? key}) : super(key: key);
+  const ColumnSortingScreen({super.key});
 
   @override
   _ColumnSortingScreenState createState() => _ColumnSortingScreenState();
