@@ -8,7 +8,7 @@ import '../../widget/pluto_example_screen.dart';
 class CopyAndPasteScreen extends StatefulWidget {
   static const routeName = 'feature/copy-and-paste';
 
-  const CopyAndPasteScreen({Key? key}) : super(key: key);
+  const CopyAndPasteScreen({super.key});
 
   @override
   _CopyAndPasteScreenState createState() => _CopyAndPasteScreenState();
