@@ -12,7 +12,7 @@ import '../../widget/pluto_example_screen.dart';
 class ExportScreen extends StatefulWidget {
   static const routeName = 'feature/export';
 
-  const ExportScreen({Key? key}) : super(key: key);
+  const ExportScreen({super.key});
 
   @override
   _ExportScreenState createState() => _ExportScreenState();
@@ -166,8 +166,7 @@ class _ExportScreenState extends State<ExportScreen> {
 class _Header extends StatefulWidget {
   const _Header({
     required this.stateManager,
-    Key? key,
-  }) : super(key: key);
+  });
 
   final PlutoGridStateManager stateManager;
 
