@@ -8,7 +8,7 @@ import '../../widget/pluto_example_screen.dart';
 class RowColorScreen extends StatefulWidget {
   static const routeName = 'feature/row-color';
 
-  const RowColorScreen({Key? key}) : super(key: key);
+  const RowColorScreen({super.key});
 
   @override
   _RowColorScreenState createState() => _RowColorScreenState();
