@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.6 from annotations
 // in pluto_grid/test/mock/shared_mocks.dart.
 // Do not manually edit this file.
 
@@ -8,12 +8,12 @@ import 'dart:ui' as _i5;
 
 import 'package:flutter/gestures.dart' as _i7;
 import 'package:flutter/material.dart' as _i3;
-import 'package:flutter/rendering.dart' as _i9;
+import 'package:flutter/rendering.dart' as _i10;
 import 'package:flutter/services.dart' as _i11;
 import 'package:flutter/src/foundation/assertions.dart' as _i12;
 import 'package:flutter/src/foundation/diagnostics.dart' as _i8;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i10;
+import 'package:mockito/src/dummies.dart' as _i9;
 import 'package:pluto_grid/pluto_grid.dart' as _i2;
 import 'package:rxdart/rxdart.dart' as _i4;
 
@@ -21,12 +21,16 @@ import 'package:rxdart/rxdart.dart' as _i4;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
+// ignore_for_file: invalid_use_of_internal_member
 
 class _FakeFilteredList_0<E> extends _i1.SmartFake
     implements _i2.FilteredList<E> {
@@ -1341,8 +1345,14 @@ class MockPlutoGridStateManager extends _i1.Mock
   @override
   String get currentSelectingText => (super.noSuchMethod(
         Invocation.getter(#currentSelectingText),
-        returnValue: '',
-        returnValueForMissingStub: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.getter(#currentSelectingText),
+        ),
+        returnValueForMissingStub: _i9.dummyValue<String>(
+          this,
+          Invocation.getter(#currentSelectingText),
+        ),
       ) as String);
 
   @override
@@ -3589,26 +3599,6 @@ class MockPlutoGridEventManager extends _i1.Mock
 class MockPlutoGridScrollController extends _i1.Mock
     implements _i2.PlutoGridScrollController {
   @override
-  set vertical(_i2.LinkedScrollControllerGroup? _vertical) =>
-      super.noSuchMethod(
-        Invocation.setter(
-          #vertical,
-          _vertical,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  set horizontal(_i2.LinkedScrollControllerGroup? _horizontal) =>
-      super.noSuchMethod(
-        Invocation.setter(
-          #horizontal,
-          _horizontal,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
   double get maxScrollHorizontal => (super.noSuchMethod(
         Invocation.getter(#maxScrollHorizontal),
         returnValue: 0.0,
@@ -3635,6 +3625,24 @@ class MockPlutoGridScrollController extends _i1.Mock
         returnValue: 0.0,
         returnValueForMissingStub: 0.0,
       ) as double);
+
+  @override
+  set vertical(_i2.LinkedScrollControllerGroup? value) => super.noSuchMethod(
+        Invocation.setter(
+          #vertical,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set horizontal(_i2.LinkedScrollControllerGroup? value) => super.noSuchMethod(
+        Invocation.setter(
+          #horizontal,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
 
   @override
   void setBodyRowsHorizontal(_i3.ScrollController? scrollController) =>
@@ -3792,18 +3800,18 @@ class MockLinkedScrollControllerGroup extends _i1.Mock
 /// See the documentation for Mockito's code generation for more information.
 class MockScrollController extends _i1.Mock implements _i3.ScrollController {
   @override
-  bool get keepScrollOffset => (super.noSuchMethod(
-        Invocation.getter(#keepScrollOffset),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-
-  @override
   double get initialScrollOffset => (super.noSuchMethod(
         Invocation.getter(#initialScrollOffset),
         returnValue: 0.0,
         returnValueForMissingStub: 0.0,
       ) as double);
+
+  @override
+  bool get keepScrollOffset => (super.noSuchMethod(
+        Invocation.getter(#keepScrollOffset),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
 
   @override
   Iterable<_i3.ScrollPosition> get positions => (super.noSuchMethod(
@@ -4015,19 +4023,6 @@ class MockScrollPosition extends _i1.Mock implements _i3.ScrollPosition {
       ) as bool);
 
   @override
-  _i3.ValueNotifier<bool> get isScrollingNotifier => (super.noSuchMethod(
-        Invocation.getter(#isScrollingNotifier),
-        returnValue: _FakeValueNotifier_25<bool>(
-          this,
-          Invocation.getter(#isScrollingNotifier),
-        ),
-        returnValueForMissingStub: _FakeValueNotifier_25<bool>(
-          this,
-          Invocation.getter(#isScrollingNotifier),
-        ),
-      ) as _i3.ValueNotifier<bool>);
-
-  @override
   double get minScrollExtent => (super.noSuchMethod(
         Invocation.getter(#minScrollExtent),
         returnValue: 0.0,
@@ -4084,11 +4079,31 @@ class MockScrollPosition extends _i1.Mock implements _i3.ScrollPosition {
       ) as bool);
 
   @override
+  bool get shouldIgnorePointer => (super.noSuchMethod(
+        Invocation.getter(#shouldIgnorePointer),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+
+  @override
   double get devicePixelRatio => (super.noSuchMethod(
         Invocation.getter(#devicePixelRatio),
         returnValue: 0.0,
         returnValueForMissingStub: 0.0,
       ) as double);
+
+  @override
+  _i3.ValueNotifier<bool> get isScrollingNotifier => (super.noSuchMethod(
+        Invocation.getter(#isScrollingNotifier),
+        returnValue: _FakeValueNotifier_25<bool>(
+          this,
+          Invocation.getter(#isScrollingNotifier),
+        ),
+        returnValueForMissingStub: _FakeValueNotifier_25<bool>(
+          this,
+          Invocation.getter(#isScrollingNotifier),
+        ),
+      ) as _i3.ValueNotifier<bool>);
 
   @override
   bool get allowImplicitScrolling => (super.noSuchMethod(
@@ -4098,11 +4113,11 @@ class MockScrollPosition extends _i1.Mock implements _i3.ScrollPosition {
       ) as bool);
 
   @override
-  _i9.ScrollDirection get userScrollDirection => (super.noSuchMethod(
+  _i10.ScrollDirection get userScrollDirection => (super.noSuchMethod(
         Invocation.getter(#userScrollDirection),
-        returnValue: _i9.ScrollDirection.idle,
-        returnValueForMissingStub: _i9.ScrollDirection.idle,
-      ) as _i9.ScrollDirection);
+        returnValue: _i10.ScrollDirection.idle,
+        returnValueForMissingStub: _i10.ScrollDirection.idle,
+      ) as _i10.ScrollDirection);
 
   @override
   bool get hasListeners => (super.noSuchMethod(
@@ -4514,7 +4529,7 @@ class MockScrollPosition extends _i1.Mock implements _i3.ScrollPosition {
       );
 
   @override
-  void didUpdateScrollDirection(_i9.ScrollDirection? direction) =>
+  void didUpdateScrollDirection(_i10.ScrollDirection? direction) =>
       super.noSuchMethod(
         Invocation.method(
           #didUpdateScrollDirection,
@@ -4716,8 +4731,8 @@ class MockStreamSubscription<T> extends _i1.Mock
           #asFuture,
           [futureValue],
         ),
-        returnValue: _i10.ifNotNull(
-              _i10.dummyValueOrNull<E>(
+        returnValue: _i9.ifNotNull(
+              _i9.dummyValueOrNull<E>(
                 this,
                 Invocation.method(
                   #asFuture,
@@ -4733,8 +4748,8 @@ class MockStreamSubscription<T> extends _i1.Mock
                 [futureValue],
               ),
             ),
-        returnValueForMissingStub: _i10.ifNotNull(
-              _i10.dummyValueOrNull<E>(
+        returnValueForMissingStub: _i9.ifNotNull(
+              _i9.dummyValueOrNull<E>(
                 this,
                 Invocation.method(
                   #asFuture,
@@ -4758,39 +4773,11 @@ class MockStreamSubscription<T> extends _i1.Mock
 /// See the documentation for Mockito's code generation for more information.
 class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
   @override
-  set onKey(_i3.FocusOnKeyCallback? _onKey) => super.noSuchMethod(
-        Invocation.setter(
-          #onKey,
-          _onKey,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  set onKeyEvent(_i3.FocusOnKeyEventCallback? _onKeyEvent) =>
-      super.noSuchMethod(
-        Invocation.setter(
-          #onKeyEvent,
-          _onKeyEvent,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
   bool get skipTraversal => (super.noSuchMethod(
         Invocation.getter(#skipTraversal),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
-  @override
-  set skipTraversal(bool? value) => super.noSuchMethod(
-        Invocation.setter(
-          #skipTraversal,
-          value,
-        ),
-        returnValueForMissingStub: null,
-      );
 
   @override
   bool get canRequestFocus => (super.noSuchMethod(
@@ -4800,15 +4787,6 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
       ) as bool);
 
   @override
-  set canRequestFocus(bool? value) => super.noSuchMethod(
-        Invocation.setter(
-          #canRequestFocus,
-          value,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
   bool get descendantsAreFocusable => (super.noSuchMethod(
         Invocation.getter(#descendantsAreFocusable),
         returnValue: false,
@@ -4816,29 +4794,11 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
       ) as bool);
 
   @override
-  set descendantsAreFocusable(bool? value) => super.noSuchMethod(
-        Invocation.setter(
-          #descendantsAreFocusable,
-          value,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
   bool get descendantsAreTraversable => (super.noSuchMethod(
         Invocation.getter(#descendantsAreTraversable),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
-  @override
-  set descendantsAreTraversable(bool? value) => super.noSuchMethod(
-        Invocation.setter(
-          #descendantsAreTraversable,
-          value,
-        ),
-        returnValueForMissingStub: null,
-      );
 
   @override
   Iterable<_i3.FocusNode> get children => (super.noSuchMethod(
@@ -4853,15 +4813,6 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
         returnValue: <_i3.FocusNode>[],
         returnValueForMissingStub: <_i3.FocusNode>[],
       ) as Iterable<_i3.FocusNode>);
-
-  @override
-  set debugLabel(String? value) => super.noSuchMethod(
-        Invocation.setter(
-          #debugLabel,
-          value,
-        ),
-        returnValueForMissingStub: null,
-      );
 
   @override
   Iterable<_i3.FocusNode> get descendants => (super.noSuchMethod(
@@ -4943,6 +4894,69 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
           Invocation.getter(#rect),
         ),
       ) as _i5.Rect);
+
+  @override
+  set skipTraversal(bool? value) => super.noSuchMethod(
+        Invocation.setter(
+          #skipTraversal,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set canRequestFocus(bool? value) => super.noSuchMethod(
+        Invocation.setter(
+          #canRequestFocus,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set descendantsAreFocusable(bool? value) => super.noSuchMethod(
+        Invocation.setter(
+          #descendantsAreFocusable,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set descendantsAreTraversable(bool? value) => super.noSuchMethod(
+        Invocation.setter(
+          #descendantsAreTraversable,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set onKey(_i3.FocusOnKeyCallback? value) => super.noSuchMethod(
+        Invocation.setter(
+          #onKey,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set onKeyEvent(_i3.FocusOnKeyEventCallback? value) => super.noSuchMethod(
+        Invocation.setter(
+          #onKeyEvent,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set debugLabel(String? value) => super.noSuchMethod(
+        Invocation.setter(
+          #debugLabel,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
 
   @override
   bool get hasListeners => (super.noSuchMethod(
@@ -5088,8 +5102,20 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
           #toStringShort,
           [],
         ),
-        returnValue: '',
-        returnValueForMissingStub: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringShort,
+            [],
+          ),
+        ),
+        returnValueForMissingStub: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringShort,
+            [],
+          ),
+        ),
       ) as String);
 
   @override
@@ -5098,7 +5124,7 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
 
   @override
   String toStringShallow({
-    String? joiner = r', ',
+    String? joiner = ', ',
     _i3.DiagnosticLevel? minLevel = _i3.DiagnosticLevel.debug,
   }) =>
       (super.noSuchMethod(
@@ -5110,15 +5136,36 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
             #minLevel: minLevel,
           },
         ),
-        returnValue: '',
-        returnValueForMissingStub: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringShallow,
+            [],
+            {
+              #joiner: joiner,
+              #minLevel: minLevel,
+            },
+          ),
+        ),
+        returnValueForMissingStub: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringShallow,
+            [],
+            {
+              #joiner: joiner,
+              #minLevel: minLevel,
+            },
+          ),
+        ),
       ) as String);
 
   @override
   String toStringDeep({
-    String? prefixLineOne = r'',
+    String? prefixLineOne = '',
     String? prefixOtherLines,
     _i3.DiagnosticLevel? minLevel = _i3.DiagnosticLevel.debug,
+    int? wrapWidth = 65,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -5128,10 +5175,35 @@ class MockFocusNode extends _i1.Mock implements _i3.FocusNode {
             #prefixLineOne: prefixLineOne,
             #prefixOtherLines: prefixOtherLines,
             #minLevel: minLevel,
+            #wrapWidth: wrapWidth,
           },
         ),
-        returnValue: '',
-        returnValueForMissingStub: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringDeep,
+            [],
+            {
+              #prefixLineOne: prefixLineOne,
+              #prefixOtherLines: prefixOtherLines,
+              #minLevel: minLevel,
+              #wrapWidth: wrapWidth,
+            },
+          ),
+        ),
+        returnValueForMissingStub: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #toStringDeep,
+            [],
+            {
+              #prefixLineOne: prefixLineOne,
+              #prefixOtherLines: prefixOtherLines,
+              #minLevel: minLevel,
+              #wrapWidth: wrapWidth,
+            },
+          ),
+        ),
       ) as String);
 
   @override
